Proper handling of 64-bit ints in PStream seed should be a 32-bit int in PRandom

+++ plearn/math/PRandom.h
@@ -89,7 +89,8 @@ protected:
     // * protected options *
     // *********************
 
-    long fixed_seed;
+    //long fixed_seed;
+    int fixed_seed;
 
 public:
 
@@ -97,7 +98,8 @@ public:
     // * public build options *
     // ************************
 
-    long seed_;
+    //long seed_; // CAUSES PROBLEMS WITH PYTHON SERVER INTERFACE
+    int seed_;
 
     // ****************
     // * Constructors *
